split recipe table into two components. adding new features to card
